Extract bridge route support helpers
