+++ src/server/codexAppServerBridge.ts
@@ -96,6 +96,12 @@ import {
   readThreadQueueState,
   writeThreadQueueState,
 } from './backendQueueProcessor.js'
+import { MethodCatalog } from './methodCatalog.js'
+import { fetchConnectorLogo, handleFileUpload, proxyTranscribe } from './mediaProxyRoutes.js'
+import { fetchCustomEndpointDefaultModel, readProviderBackedModelIds } from './providerModelDiscovery.js'
+import { cloneGithubRepositoryIntoBase, createProjectlessThreadDirectory } from './projectCreation.js'
+import { API_PERF_BODY_MB_THRESHOLD, API_PERF_LOGGING_ENABLED, API_PERF_MS_THRESHOLD, getChunkByteLength } from './apiPerfConfig.js'
+import { buildThreadSearchIndex, isExactPhraseMatch, listFilesWithRipgrep, scoreFileCandidate, type ThreadSearchIndex } from './threadSearchIndex.js'
 import { mergeSessionSkillInputsIntoThreadResult, sanitizeThreadTurnsInlinePayloads } from './threadInlinePayloads.js'
 export { mergeSessionSkillInputsIntoTurns, sanitizeThreadTurnsInlinePayloads } from './threadInlinePayloads.js'
 
@@ -165,117 +171,12 @@ type ChatgptAuthTokensRefreshResponse = {
   chatgptPlanType: string | null
 }
 
-type ThreadSearchDocument = {
-  id: string
-  title: string
-  preview: string
-  messageText: string
-  searchableText: string
-}
-
-type ThreadSearchIndex = {
-  docsById: Map<string, ThreadSearchDocument>
-}
-
-type ProviderModelsResponse = {
-  data: string[]
-  providerId: string
-  source: 'provider'
-}
-
-const PROVIDER_MODELS_FETCH_TIMEOUT_MS = 5_000
-
 const THREAD_RESPONSE_TURN_LIMIT = 10
 const THREAD_TURN_PAGE_READ_CACHE_TTL_MS = 30_000
 const THREAD_METHODS_WITH_TURNS = new Set(['thread/read', 'thread/resume', 'thread/fork', 'thread/rollback'])
 const THREAD_SEARCH_FULL_TEXT_THREAD_LIMIT = 100
 const PROJECTLESS_THREAD_DIRECTORY_MAX_ATTEMPTS = 100
 const PROJECTLESS_THREAD_SLUG_MAX_LENGTH = 80
-const API_PERF_LOGGING_ENV_KEY = 'CODEXUI_API_PERF_LOGGING'
-const API_PERF_MS_THRESHOLD_ENV_KEY = 'CODEXUI_API_PERF_MS_THRESHOLD'
-const API_PERF_BODY_MB_THRESHOLD_ENV_KEY = 'CODEXUI_API_PERF_BODY_MB_THRESHOLD'
-const DEFAULT_API_PERF_MS_THRESHOLD = 300
-const DEFAULT_API_PERF_BODY_MB_THRESHOLD = 1
-const MB_DIVISOR = 1024 * 1024
-function readEnvValueFromFile(filePath: string, key: string): string | null {
-  try {
-    const content = readFileSync(filePath, 'utf8')
-    const escapedKey = key.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')
-    const match = content.match(new RegExp(`^\\s*${escapedKey}\\s*=\\s*(.+)\\s*$`, 'm'))
-    if (!match) return null
-    const rawValue = match[1]?.trim() ?? ''
-    if (!rawValue) return null
-    if ((rawValue.startsWith('"') && rawValue.endsWith('"')) || (rawValue.startsWith('\'') && rawValue.endsWith('\''))) {
-      return rawValue.slice(1, -1).trim()
-    }
-    return rawValue
-  } catch {
-    return null
-  }
-}
-
-function parseBooleanEnvFlag(value: string | null | undefined): boolean | null {
-  if (!value) return null
-  const normalized = value.trim().toLowerCase()
-  if (['1', 'true', 'yes', 'on'].includes(normalized)) return true
-  if (['0', 'false', 'no', 'off'].includes(normalized)) return false
-  return null
-}
-
-function resolveApiPerfLoggingEnabled(): boolean {
-  const explicitValue = parseBooleanEnvFlag(process.env[API_PERF_LOGGING_ENV_KEY])
-  if (explicitValue !== null) return explicitValue
-
-  const fromEnvLocal = parseBooleanEnvFlag(readEnvValueFromFile('.env.local', API_PERF_LOGGING_ENV_KEY))
-  if (fromEnvLocal !== null) return fromEnvLocal
-
-  const fromEnv = parseBooleanEnvFlag(readEnvValueFromFile('.env', API_PERF_LOGGING_ENV_KEY))
-  if (fromEnv !== null) return fromEnv
-
-  return false
-}
-
-const API_PERF_LOGGING_ENABLED = resolveApiPerfLoggingEnabled()
-
-function parseNumberEnvFlag(value: string | null | undefined): number | null {
-  if (!value) return null
-  const parsed = Number.parseFloat(value.trim())
-  if (!Number.isFinite(parsed)) return null
-  return parsed
-}
-
-function resolveNumericEnvConfig(envKey: string, fallback: number): number {
-  const fromProcess = parseNumberEnvFlag(process.env[envKey])
-  if (fromProcess !== null) return fromProcess
-
-  const fromEnvLocal = parseNumberEnvFlag(readEnvValueFromFile('.env.local', envKey))
-  if (fromEnvLocal !== null) return fromEnvLocal
-
-  const fromEnv = parseNumberEnvFlag(readEnvValueFromFile('.env', envKey))
-  if (fromEnv !== null) return fromEnv
-
-  return fallback
-}
-
-const API_PERF_MS_THRESHOLD = resolveNumericEnvConfig(API_PERF_MS_THRESHOLD_ENV_KEY, DEFAULT_API_PERF_MS_THRESHOLD)
-const API_PERF_BODY_MB_THRESHOLD = resolveNumericEnvConfig(API_PERF_BODY_MB_THRESHOLD_ENV_KEY, DEFAULT_API_PERF_BODY_MB_THRESHOLD)
-
-function getChunkByteLength(chunk: unknown, encoding?: BufferEncoding): number {
-  if (typeof chunk === 'string') {
-    return Buffer.byteLength(chunk, encoding)
-  }
-  if (chunk instanceof Uint8Array) {
-    return chunk.byteLength
-  }
-  if (ArrayBuffer.isView(chunk)) {
-    return chunk.byteLength
-  }
-  if (chunk instanceof ArrayBuffer) {
-    return chunk.byteLength
-  }
-  return 0
-}
-
 function asRecord(value: unknown): Record<string, unknown> | null {
   return value !== null && typeof value === 'object' && !Array.isArray(value)
     ? (value as Record<string, unknown>)
@@ -332,333 +233,6 @@ function logProviderModelDiscoveryWarning(message: string, details: Record<strin
 
 function isTimeoutError(payload: unknown): boolean {
   return payload instanceof Error && (payload.name === 'AbortError' || payload.name === 'TimeoutError')
-}
-
-function formatProjectlessDateSegment(date = new Date()): string {
-  const month = String(date.getMonth() + 1).padStart(2, '0')
-  const day = String(date.getDate()).padStart(2, '0')
-  return `${date.getFullYear()}-${month}-${day}`
-}
-
-function buildProjectlessPromptSlug(prompt: string | null): string {
-  const slug = prompt
-    ?.toLowerCase()
-    .match(/[a-z0-9]+/g)
-    ?.slice(0, 6)
-    .join('-')
-    .slice(0, PROJECTLESS_THREAD_SLUG_MAX_LENGTH)
-  return slug && slug.length > 0 ? slug : 'new-chat'
-}
-
-async function ensureRealDirectory(path: string, label: string): Promise<void> {
-  const info = await lstat(path)
-  if (info.isSymbolicLink() || !info.isDirectory()) {
-    throw new Error(`${label} must be a real directory`)
-  }
-}
-
-async function createProjectlessThreadDirectory(prompt: string | null): Promise<{ cwd: string; outputDirectory: string; workspaceRoot: string }> {
-  const workspaceRoot = join(homedir(), 'Documents', 'Codex')
-  await mkdir(workspaceRoot, { recursive: true })
-  await ensureRealDirectory(workspaceRoot, 'Projectless workspace root')
-
-  const dateDir = join(workspaceRoot, formatProjectlessDateSegment())
-  await mkdir(dateDir, { recursive: true })
-  await ensureRealDirectory(dateDir, 'Projectless thread date directory')
-
-  const slug = buildProjectlessPromptSlug(prompt)
-  for (let index = 0; index < PROJECTLESS_THREAD_DIRECTORY_MAX_ATTEMPTS; index += 1) {
-    const folderName = index === 0 ? slug : `${slug}-${index + 1}`
-    const cwd = join(dateDir, folderName)
-    try {
-      await mkdir(cwd, { recursive: false })
-      return { cwd, outputDirectory: cwd, workspaceRoot }
-    } catch {
-      try {
-        await stat(cwd)
-      } catch {
-        throw new Error('Failed to create new chat folder')
-      }
-    }
-  }
-
-  throw new Error('Unable to create a unique new chat folder')
-}
-
-function normalizeGithubCloneUrl(rawUrl: string): { url: string; repoName: string } {
-  const trimmedUrl = rawUrl.trim()
-  if (!trimmedUrl) throw new Error('Missing GitHub repository URL')
-
-  const sshMatch = trimmedUrl.match(/^git@github\.com:([A-Za-z0-9_.-]+)\/([A-Za-z0-9_.-]+?)(?:\.git)?$/u)
-  if (sshMatch) {
-    const repoName = sshMatch[2]
-    return { url: `git@github.com:${sshMatch[1]}/${repoName}.git`, repoName }
-  }
-
-  let parsed: URL
-  try {
-    parsed = new URL(trimmedUrl)
-  } catch {
-    throw new Error('Enter a valid GitHub repository URL')
-  }
-  if (parsed.hostname.toLowerCase() !== 'github.com') {
-    throw new Error('Only github.com repository URLs are supported')
-  }
-  const segments = parsed.pathname.split('/').filter(Boolean)
-  if (segments.length < 2) {
-    throw new Error('Enter a GitHub repository URL with owner and repository name')
-  }
-  const owner = segments[0]
-  const repoName = segments[1].replace(/\.git$/iu, '')
-  if (!/^[A-Za-z0-9_.-]+$/u.test(owner) || !/^[A-Za-z0-9_.-]+$/u.test(repoName)) {
-    throw new Error('GitHub repository owner or name contains unsupported characters')
-  }
-  return { url: `https://github.com/${owner}/${repoName}.git`, repoName }
-}
-
-async function cloneGithubRepositoryIntoBase(rawUrl: string, rawBasePath: string): Promise<string> {
-  const basePath = rawBasePath.trim()
-  if (!basePath) throw new Error('Missing clone destination folder')
-  const normalizedBasePath = isAbsolute(basePath) ? basePath : resolve(basePath)
-  await ensureRealDirectory(normalizedBasePath, 'Clone destination folder')
-
-  const { url, repoName } = normalizeGithubCloneUrl(rawUrl)
-  const targetPath = join(normalizedBasePath, repoName)
-  try {
-    await stat(targetPath)
-    throw new Error(`Destination already exists: ${targetPath}`)
-  } catch (error) {
-    if ((error as NodeJS.ErrnoException)?.code !== 'ENOENT') throw error
-  }
-
-  try {
-    await runCommand('git', ['clone', url, targetPath], { cwd: normalizedBasePath, timeoutMs: 5 * 60_000 })
-  } catch (error) {
-    await rm(targetPath, { recursive: true, force: true }).catch(() => undefined)
-    throw error
-  }
-  await persistWorkspaceRoot(targetPath, '')
-  return targetPath
-}
-
-function normalizeHeaderValue(value: unknown): string | null {
-  if (typeof value === 'string') {
-    const trimmed = value.trim()
-    return trimmed.length > 0 ? trimmed : null
-  }
-  if (typeof value === 'number' || typeof value === 'boolean') {
-    return String(value)
-  }
-  return null
-}
-
-function normalizeQueryParams(value: unknown): URLSearchParams {
-  const params = new URLSearchParams()
-  const record = asRecord(value)
-  if (!record) return params
-
-  for (const [key, rawValue] of Object.entries(record)) {
-    const normalized = normalizeHeaderValue(rawValue)
-    if (!normalized) continue
-    params.set(key, normalized)
-  }
-
-  return params
-}
-
-function buildProviderModelsUrl(baseUrl: string, queryParams: unknown): URL {
-  const url = new URL(baseUrl)
-  url.pathname = url.pathname.endsWith('/') ? `${url.pathname}models` : `${url.pathname}/models`
-  const extraParams = normalizeQueryParams(queryParams)
-  for (const [key, value] of extraParams.entries()) {
-    url.searchParams.set(key, value)
-  }
-  return url
-}
-
-function normalizeProviderModelsData(payload: unknown): string[] {
-  const record = asRecord(payload)
-  const rows = Array.isArray(record?.data) ? record.data : null
-  if (!rows) {
-    throw new Error('provider /models payload is missing a data array')
-  }
-
-  const ids: string[] = []
-  for (const row of rows) {
-    const entry = asRecord(row)
-    const candidate = readNonEmptyString(entry?.id)
-    if (!candidate || ids.includes(candidate)) continue
-    ids.push(candidate)
-  }
-  return ids
-}
-
-async function fetchCustomEndpointDefaultModel(baseUrl: string, apiKey: string): Promise<string> {
-  const normalizedBaseUrl = baseUrl.trim()
-  if (!normalizedBaseUrl) return ''
-
-  try {
-    const modelsUrl = buildProviderModelsUrl(normalizedBaseUrl, null)
-    const headers: Record<string, string> = apiKey ? { Authorization: `Bearer ${apiKey}` } : {}
-    const response = await fetch(modelsUrl, { headers, signal: AbortSignal.timeout(PROVIDER_MODELS_FETCH_TIMEOUT_MS) })
-    if (!response.ok) return ''
-    const payload = await response.json() as unknown
-    const modelIds = normalizeProviderModelsData(payload)
-    return modelIds[0] ?? ''
-  } catch {
-    return ''
-  }
-}
-
-async function readProviderBackedModelIds(appServer: AppServerProcess): Promise<ProviderModelsResponse> {
-  const configPayload = asRecord(await appServer.rpc('config/read', {}))
-  const config = asRecord(configPayload?.config)
-  const providerId = readNonEmptyString(config?.model_provider)
-  if (!providerId) {
-    return { data: [], providerId: '', source: 'provider' }
-  }
-
-  const providers = asRecord(config?.model_providers)
-  const provider = asRecord(providers?.[providerId])
-  if (!provider) {
-    logProviderModelDiscoveryWarning('configured provider is missing from model_providers', { providerId })
-    return { data: [], providerId, source: 'provider' }
-  }
-
-  const wireApi = readNonEmptyString(provider.wire_api)
-  if (wireApi !== 'responses') {
-    return { data: [], providerId, source: 'provider' }
-  }
-
-  const baseUrl = readNonEmptyString(provider.base_url)
-  if (!baseUrl) {
-    logProviderModelDiscoveryWarning('responses provider is missing base_url', { providerId })
-    return { data: [], providerId, source: 'provider' }
-  }
-
-  const headers = new Headers()
-  const configuredHeaders = asRecord(provider.http_headers)
-  if (configuredHeaders) {
-    for (const [key, rawValue] of Object.entries(configuredHeaders)) {
-      const normalized = normalizeHeaderValue(rawValue)
-      if (!normalized) continue
-      headers.set(key, normalized)
-    }
-  }
-
-  const bearerToken = readNonEmptyString(provider.experimental_bearer_token)
-  if (bearerToken && !headers.has('Authorization')) {
-    headers.set('Authorization', `Bearer ${bearerToken}`)
-  }
-
-  const envKey = readNonEmptyString(provider.env_key)
-  const envHttpHeaders = asRecord(provider.env_http_headers)
-  if (envKey || envHttpHeaders) {
-    logProviderModelDiscoveryWarning('provider discovery skipped env-backed auth/header expansion', {
-      providerId,
-      hasEnvKey: Boolean(envKey),
-      hasEnvHttpHeaders: Boolean(envHttpHeaders),
-    })
-  }
-
-  let requestUrl: URL
-  try {
-    requestUrl = buildProviderModelsUrl(baseUrl, provider.query_params)
-  } catch (error) {
-    logProviderModelDiscoveryWarning('provider /models URL was invalid', {
-      providerId,
-      error: getErrorMessage(error, 'invalid url'),
-    })
-    return { data: [], providerId, source: 'provider' }
-  }
-
-  let response: Response
-  try {
-    response = await fetch(requestUrl, {
-      method: 'GET',
-      headers,
-      signal: AbortSignal.timeout(PROVIDER_MODELS_FETCH_TIMEOUT_MS),
-    })
-  } catch (error) {
-    logProviderModelDiscoveryWarning('provider /models request failed', {
-      providerId,
-      error: isTimeoutError(error) ? `request timed out after ${PROVIDER_MODELS_FETCH_TIMEOUT_MS}ms` : getErrorMessage(error, 'network error'),
-    })
-    return { data: [], providerId, source: 'provider' }
-  }
-
-  let payload: unknown = null
-  try {
-    payload = await response.json()
-  } catch (error) {
-    logProviderModelDiscoveryWarning('provider /models response was not valid JSON', {
-      providerId,
-      status: response.status,
-      error: getErrorMessage(error, 'invalid json'),
-    })
-    return { data: [], providerId, source: 'provider' }
-  }
-
-  if (!response.ok) {
-    logProviderModelDiscoveryWarning('provider /models request returned non-2xx', {
-      providerId,
-      status: response.status,
-      statusText: response.statusText,
-    })
-    return { data: [], providerId, source: 'provider' }
-  }
-
-  try {
-    return {
-      data: normalizeProviderModelsData(payload),
-      providerId,
-      source: 'provider',
-    }
-  } catch (error) {
-    logProviderModelDiscoveryWarning('provider /models payload was invalid', {
-      providerId,
-      error: getErrorMessage(error, 'invalid payload'),
-    })
-    return { data: [], providerId, source: 'provider' }
-  }
-}
-
-function extractThreadMessageText(threadReadPayload: unknown): string {
-  const payload = asRecord(threadReadPayload)
-  const thread = asRecord(payload?.thread)
-  const turns = Array.isArray(thread?.turns) ? thread.turns : []
-  const parts: string[] = []
-
-  for (const turn of turns) {
-    const turnRecord = asRecord(turn)
-    const items = Array.isArray(turnRecord?.items) ? turnRecord.items : []
-    for (const item of items) {
-      const itemRecord = asRecord(item)
-      const type = typeof itemRecord?.type === 'string' ? itemRecord.type : ''
-      if (type === 'agentMessage' && typeof itemRecord?.text === 'string' && itemRecord.text.trim().length > 0) {
-        parts.push(itemRecord.text.trim())
-        continue
-      }
-      if (type === 'userMessage') {
-        const content = Array.isArray(itemRecord?.content) ? itemRecord.content : []
-        for (const block of content) {
-          const blockRecord = asRecord(block)
-          if (blockRecord?.type === 'text' && typeof blockRecord.text === 'string' && blockRecord.text.trim().length > 0) {
-            parts.push(blockRecord.text.trim())
-          }
-        }
-        continue
-      }
-      if (type === 'commandExecution') {
-        const command = typeof itemRecord?.command === 'string' ? itemRecord.command.trim() : ''
-        const output = typeof itemRecord?.aggregatedOutput === 'string' ? itemRecord.aggregatedOutput.trim() : ''
-        if (command) parts.push(command)
-        if (output) parts.push(output)
-      }
-    }
-  }
-
-  return parts.join('\n').trim()
 }
 
 function readNonEmptyString(value: unknown): string {
@@ -721,62 +295,6 @@ export async function callRpcWithArchiveRecovery(
     })
     return appServer.rpc(method, params ?? null)
   }
-}
-
-function isExactPhraseMatch(query: string, doc: ThreadSearchDocument): boolean {
-  const q = query.trim().toLowerCase()
-  if (!q) return false
-  return (
-    doc.title.toLowerCase().includes(q) ||
-    doc.preview.toLowerCase().includes(q) ||
-    doc.messageText.toLowerCase().includes(q)
-  )
-}
-
-function scoreFileCandidate(path: string, query: string): number {
-  if (!query) return 0
-  const lowerPath = path.toLowerCase()
-  const lowerQuery = query.toLowerCase()
-  const baseName = lowerPath.slice(lowerPath.lastIndexOf('/') + 1)
-  if (baseName === lowerQuery) return 0
-  if (baseName.startsWith(lowerQuery)) return 1
-  if (baseName.includes(lowerQuery)) return 2
-  if (lowerPath.includes(`/${lowerQuery}`)) return 3
-  if (lowerPath.includes(lowerQuery)) return 4
-  return 10
-}
-
-async function listFilesWithRipgrep(cwd: string): Promise<string[]> {
-  return await new Promise<string[]>((resolve, reject) => {
-    const ripgrepCommand = resolveRipgrepCommand()
-    if (!ripgrepCommand) {
-      reject(new Error('ripgrep (rg) is not available'))
-      return
-    }
-
-    const proc = spawn(ripgrepCommand, ['--files', '--hidden', '-g', '!.git', '-g', '!node_modules'], {
-      cwd,
-      env: process.env,
-      stdio: ['ignore', 'pipe', 'pipe'],
-    })
-    let stdout = ''
-    let stderr = ''
-    proc.stdout.on('data', (chunk: Buffer) => { stdout += chunk.toString() })
-    proc.stderr.on('data', (chunk: Buffer) => { stderr += chunk.toString() })
-    proc.on('error', reject)
-    proc.on('close', (code) => {
-      if (code === 0) {
-        const rows = stdout
-          .split(/\r?\n/)
-          .map((line) => line.trim())
-          .filter(Boolean)
-        resolve(rows)
-        return
-      }
-      const details = [stderr.trim(), stdout.trim()].filter(Boolean).join('\n')
-      reject(new Error(details || 'rg --files failed'))
-    })
-  })
 }
 
 function getCodexHomeDir(): string {
@@ -1458,319 +976,6 @@ async function readRawBody(req: IncomingMessage): Promise<Buffer> {
   return Buffer.concat(chunks)
 }
 
-function bufferIndexOf(buf: Buffer, needle: Buffer, start = 0): number {
-  for (let i = start; i <= buf.length - needle.length; i++) {
-    let match = true
-    for (let j = 0; j < needle.length; j++) {
-      if (buf[i + j] !== needle[j]) { match = false; break }
-    }
-    if (match) return i
-  }
-  return -1
-}
-
-function handleFileUpload(req: IncomingMessage, res: ServerResponse): void {
-  const chunks: Buffer[] = []
-  req.on('data', (chunk: Buffer) => chunks.push(chunk))
-  req.on('end', async () => {
-    try {
-      const body = Buffer.concat(chunks)
-      const contentType = req.headers['content-type'] ?? ''
-      const boundaryMatch = contentType.match(/boundary=(.+)/i)
-      if (!boundaryMatch) { setJson(res, 400, { error: 'Missing multipart boundary' }); return }
-      const boundary = boundaryMatch[1]
-      const boundaryBuf = Buffer.from(`--${boundary}`)
-      const parts: Buffer[] = []
-      let searchStart = 0
-      while (searchStart < body.length) {
-        const idx = body.indexOf(boundaryBuf, searchStart)
-        if (idx < 0) break
-        if (searchStart > 0) parts.push(body.subarray(searchStart, idx))
-        searchStart = idx + boundaryBuf.length
-        if (body[searchStart] === 0x0d && body[searchStart + 1] === 0x0a) searchStart += 2
-      }
-      let fileName = 'uploaded-file'
-      let fileData: Buffer | null = null
-      const headerSep = Buffer.from('\r\n\r\n')
-      for (const part of parts) {
-        const headerEnd = bufferIndexOf(part, headerSep)
-        if (headerEnd < 0) continue
-        const headers = part.subarray(0, headerEnd).toString('utf8')
-        const fnMatch = headers.match(/filename="([^"]+)"/i)
-        if (!fnMatch) continue
-        fileName = fnMatch[1].replace(/[/\\]/g, '_')
-        let end = part.length
-        if (end >= 2 && part[end - 2] === 0x0d && part[end - 1] === 0x0a) end -= 2
-        fileData = part.subarray(headerEnd + 4, end)
-        break
-      }
-      if (!fileData) { setJson(res, 400, { error: 'No file in request' }); return }
-      const uploadDir = join(tmpdir(), 'codex-web-uploads')
-      await mkdir(uploadDir, { recursive: true })
-      const destDir = await mkdtemp(join(uploadDir, 'f-'))
-      const destPath = join(destDir, fileName)
-      await writeFile(destPath, fileData)
-      setJson(res, 200, { path: destPath })
-    } catch (err) {
-      setJson(res, 500, { error: getErrorMessage(err, 'Upload failed') })
-    }
-  })
-  req.on('error', (err: Error) => {
-    setJson(res, 500, { error: getErrorMessage(err, 'Upload stream error') })
-  })
-}
-
-function httpPost(
-  url: string,
-  headers: Record<string, string | number>,
-  body: Buffer,
-): Promise<{ status: number; body: string }> {
-  const doRequest = url.startsWith('http://') ? httpRequest : httpsRequest
-  return new Promise((resolve, reject) => {
-    const req = doRequest(url, { method: 'POST', headers }, (res) => {
-      const chunks: Buffer[] = []
-      res.on('data', (c: Buffer) => chunks.push(c))
-      res.on('end', () => resolve({ status: res.statusCode ?? 500, body: Buffer.concat(chunks).toString('utf8') }))
-      res.on('error', reject)
-    })
-    req.on('error', reject)
-    req.write(body)
-    req.end()
-  })
-}
-
-let curlImpersonateAvailable: boolean | null = null
-
-function curlImpersonatePost(
-  url: string,
-  headers: Record<string, string | number>,
-  body: Buffer,
-): Promise<{ status: number; body: string }> {
-  return new Promise((resolve, reject) => {
-    const args = ['-s', '-w', '\n%{http_code}', '-X', 'POST', url]
-    for (const [k, v] of Object.entries(headers)) {
-      if (k.toLowerCase() === 'content-length') continue
-      args.push('-H', `${k}: ${String(v)}`)
-    }
-    args.push('--data-binary', '@-')
-    const proc = spawn('curl-impersonate-chrome', args, {
-      env: { ...process.env, CURL_IMPERSONATE: 'chrome116' },
-      stdio: ['pipe', 'pipe', 'pipe'],
-    })
-    const chunks: Buffer[] = []
-    proc.stdout.on('data', (c: Buffer) => chunks.push(c))
-    proc.on('error', (e) => {
-      curlImpersonateAvailable = false
-      reject(e)
-    })
-    proc.on('close', (code) => {
-      const raw = Buffer.concat(chunks).toString('utf8')
-      const lastNewline = raw.lastIndexOf('\n')
-      const statusStr = lastNewline >= 0 ? raw.slice(lastNewline + 1).trim() : ''
-      const responseBody = lastNewline >= 0 ? raw.slice(0, lastNewline) : raw
-      const status = parseInt(statusStr, 10) || (code === 0 ? 200 : 500)
-      curlImpersonateAvailable = true
-      resolve({ status, body: responseBody })
-    })
-    proc.stdin.write(body)
-    proc.stdin.end()
-  })
-}
-
-async function proxyTranscribe(
-  body: Buffer,
-  contentType: string,
-  authToken: string,
-  accountId?: string,
-): Promise<{ status: number; body: string }> {
-  const chatgptHeaders: Record<string, string | number> = {
-    'Content-Type': contentType,
-    'Content-Length': body.length,
-    Authorization: `Bearer ${authToken}`,
-    originator: 'Codex Desktop',
-    'User-Agent': `Codex Desktop/0.1.0 (${process.platform}; ${process.arch})`,
-  }
-  if (accountId) chatgptHeaders['ChatGPT-Account-Id'] = accountId
-
-  const postFn = curlImpersonateAvailable !== false ? curlImpersonatePost : httpPost
-  let result: { status: number; body: string }
-  try {
-    result = await postFn('https://chatgpt.com/backend-api/transcribe', chatgptHeaders, body)
-  } catch {
-    result = await httpPost('https://chatgpt.com/backend-api/transcribe', chatgptHeaders, body)
-  }
-
-  if (result.status === 403 && result.body.includes('cf_chl')) {
-    if (curlImpersonateAvailable !== false && postFn !== curlImpersonatePost) {
-      try {
-        const ciResult = await curlImpersonatePost('https://chatgpt.com/backend-api/transcribe', chatgptHeaders, body)
-        if (ciResult.status !== 403) return ciResult
-      } catch {}
-    }
-    return { status: 503, body: JSON.stringify({ error: 'Transcription blocked by Cloudflare. Install curl-impersonate-chrome.' }) }
-  }
-
-  return result
-}
-
-function parseConnectorLogoUrl(rawUrl: string): { connectorId: string; theme: 'light' | 'dark' } | null {
-  const trimmed = rawUrl.trim()
-  if (!trimmed.startsWith('connectors://')) return null
-  const rest = trimmed.slice('connectors://'.length)
-  const connectorId = (rest.split(/[/?#]/u)[0] ?? '').trim()
-  if (!connectorId) return null
-  const query = rest.includes('?') ? rest.slice(rest.indexOf('?') + 1).split('#')[0] ?? '' : ''
-  const theme = new URLSearchParams(query).get('theme')?.toLowerCase() === 'dark' ? 'dark' : 'light'
-  return { connectorId, theme }
-}
-
-async function fetchConnectorLogo(rawUrl: string): Promise<{ contentType: string; body: Buffer }> {
-  const parsed = parseConnectorLogoUrl(rawUrl)
-  if (!parsed) throw new Error('Unsupported connector logo URL')
-  const auth = await readCodexAuth()
-  if (!auth) throw new Error('No auth token available for connector logo')
-
-  const endpoint = `https://chatgpt.com/backend-api/aip/connectors/${encodeURIComponent(parsed.connectorId)}/logo?theme=${parsed.theme}`
-  const response = await fetch(endpoint, {
-    headers: {
-      Authorization: `Bearer ${auth.accessToken}`,
-      originator: 'Codex Desktop',
-      'User-Agent': `Codex Desktop/0.1.0 (${process.platform}; ${process.arch})`,
-      ...(auth.accountId ? { 'ChatGPT-Account-Id': auth.accountId } : {}),
-    },
-    signal: AbortSignal.timeout(10_000),
-  })
-  if (!response.ok) throw new Error(`Connector logo fetch failed (${response.status})`)
-
-  const contentType = response.headers.get('content-type') ?? ''
-  if (contentType.includes('application/json')) {
-    const payload = asRecord(await response.json())
-    const body = asRecord(payload?.body)
-    const base64 = readNonEmptyString(body?.base64)
-    const nestedContentType = readNonEmptyString(body?.contentType) ?? readNonEmptyString(body?.content_type)
-    if (!base64 || !nestedContentType) throw new Error('Connector logo response was missing image data')
-    return { contentType: nestedContentType, body: Buffer.from(base64, 'base64') }
-  }
-
-  return {
-    contentType: contentType || 'image/png',
-    body: Buffer.from(await response.arrayBuffer()),
-  }
-}
-
-class MethodCatalog {
-  private methodCache: string[] | null = null
-  private notificationCache: string[] | null = null
-
-  private async runGenerateSchemaCommand(outDir: string): Promise<void> {
-    await new Promise<void>((resolve, reject) => {
-      const codexCommand = resolveCodexCommand()
-      if (!codexCommand) {
-        reject(new Error('Codex CLI is not available. Install @openai/codex or set CODEXUI_CODEX_COMMAND.'))
-        return
-      }
-
-      const invocation = getSpawnInvocation(codexCommand, ['app-server', 'generate-json-schema', '--out', outDir])
-      const process = spawn(invocation.command, invocation.args, {
-        stdio: ['ignore', 'ignore', 'pipe'],
-      })
-
-      let stderr = ''
-
-      process.stderr.setEncoding('utf8')
-      process.stderr.on('data', (chunk: string) => {
-        stderr += chunk
-      })
-
-      process.on('error', reject)
-      process.on('exit', (code) => {
-        if (code === 0) {
-          resolve()
-          return
-        }
-
-        reject(new Error(stderr.trim() || `generate-json-schema exited with code ${String(code)}`))
-      })
-    })
-  }
-
-  private extractMethodsFromClientRequest(payload: unknown): string[] {
-    const root = asRecord(payload)
-    const oneOf = Array.isArray(root?.oneOf) ? root.oneOf : []
-    const methods = new Set<string>()
-
-    for (const entry of oneOf) {
-      const row = asRecord(entry)
-      const properties = asRecord(row?.properties)
-      const methodDef = asRecord(properties?.method)
-      const methodEnum = Array.isArray(methodDef?.enum) ? methodDef.enum : []
-
-      for (const item of methodEnum) {
-        if (typeof item === 'string' && item.length > 0) {
-          methods.add(item)
-        }
-      }
-    }
-
-    return Array.from(methods).sort((a, b) => a.localeCompare(b))
-  }
-
-  private extractMethodsFromServerNotification(payload: unknown): string[] {
-    const root = asRecord(payload)
-    const oneOf = Array.isArray(root?.oneOf) ? root.oneOf : []
-    const methods = new Set<string>()
-
-    for (const entry of oneOf) {
-      const row = asRecord(entry)
-      const properties = asRecord(row?.properties)
-      const methodDef = asRecord(properties?.method)
-      const methodEnum = Array.isArray(methodDef?.enum) ? methodDef.enum : []
-
-      for (const item of methodEnum) {
-        if (typeof item === 'string' && item.length > 0) {
-          methods.add(item)
-        }
-      }
-    }
-
-    return Array.from(methods).sort((a, b) => a.localeCompare(b))
-  }
-
-  async listMethods(): Promise<string[]> {
-    if (this.methodCache) {
-      return this.methodCache
-    }
-
-    const outDir = await mkdtemp(join(tmpdir(), 'codex-web-local-schema-'))
-    await this.runGenerateSchemaCommand(outDir)
-
-    const clientRequestPath = join(outDir, 'ClientRequest.json')
-    const raw = await readFile(clientRequestPath, 'utf8')
-    const parsed = JSON.parse(raw) as unknown
-    const methods = this.extractMethodsFromClientRequest(parsed)
-
-    this.methodCache = methods
-    return methods
-  }
-
-  async listNotificationMethods(): Promise<string[]> {
-    if (this.notificationCache) {
-      return this.notificationCache
-    }
-
-    const outDir = await mkdtemp(join(tmpdir(), 'codex-web-local-schema-'))
-    await this.runGenerateSchemaCommand(outDir)
-
-    const serverNotificationPath = join(outDir, 'ServerNotification.json')
-    const raw = await readFile(serverNotificationPath, 'utf8')
-    const parsed = JSON.parse(raw) as unknown
-    const methods = this.extractMethodsFromServerNotification(parsed)
-
-    this.notificationCache = methods
-    return methods
-  }
-}
-
 type CodexBridgeMiddleware = ((req: IncomingMessage, res: ServerResponse, next: () => void) => Promise<void>) & {
   dispose: () => void
   subscribeNotifications: (listener: (value: { method: string; params: unknown; atIso: string }) => void) => () => void
@@ -1820,82 +1025,6 @@ function getSharedBridgeState(): SharedBridgeState {
   }
   globalScope[SHARED_BRIDGE_KEY] = created
   return created
-}
-
-async function loadAllThreadsForSearch(appServer: AppServerProcess): Promise<ThreadSearchDocument[]> {
-  const threads: Array<{ id: string; title: string; preview: string }> = []
-  let cursor: string | null = null
-
-  do {
-    const response = asRecord(await appServer.rpc('thread/list', {
-      archived: false,
-      limit: 100,
-      sortKey: 'updated_at',
-      modelProviders: [],
-      cursor,
-    }))
-    const data = Array.isArray(response?.data) ? response.data : []
-    for (const row of data) {
-      const record = asRecord(row)
-      const id = typeof record?.id === 'string' ? record.id : ''
-      if (!id) continue
-      const title = typeof record?.name === 'string' && record.name.trim().length > 0
-        ? record.name.trim()
-        : (typeof record?.preview === 'string' && record.preview.trim().length > 0 ? record.preview.trim() : 'Untitled thread')
-      const preview = typeof record?.preview === 'string' ? record.preview : ''
-      threads.push({ id, title, preview })
-    }
-    cursor = typeof response?.nextCursor === 'string' && response.nextCursor.length > 0 ? response.nextCursor : null
-  } while (cursor)
-
-  const docs: ThreadSearchDocument[] = threads.map((thread) => {
-    const searchableText = [thread.title, thread.preview].filter(Boolean).join('\n')
-    return {
-      id: thread.id,
-      title: thread.title,
-      preview: thread.preview,
-      messageText: '',
-      searchableText,
-    } satisfies ThreadSearchDocument
-  })
-
-  const docsById = new Map<string, ThreadSearchDocument>(docs.map((doc) => [doc.id, doc]))
-  const fullTextThreads = threads.slice(0, THREAD_SEARCH_FULL_TEXT_THREAD_LIMIT)
-  const concurrency = 4
-  for (let offset = 0; offset < fullTextThreads.length; offset += concurrency) {
-    const batch = fullTextThreads.slice(offset, offset + concurrency)
-    const loaded = await Promise.all(batch.map(async (thread) => {
-      try {
-        const readResponse = await appServer.rpc('thread/read', {
-          threadId: thread.id,
-          includeTurns: true,
-        })
-        const messageText = extractThreadMessageText(readResponse)
-        const searchableText = [thread.title, thread.preview, messageText].filter(Boolean).join('\n')
-        return [thread.id, {
-          id: thread.id,
-          title: thread.title,
-          preview: thread.preview,
-          messageText,
-          searchableText,
-        } satisfies ThreadSearchDocument] as const
-      } catch {
-        return null
-      }
-    }))
-    for (const row of loaded) {
-      if (!row) continue
-      docsById.set(row[0], row[1])
-    }
-  }
-
-  return Array.from(docsById.values())
-}
-
-async function buildThreadSearchIndex(appServer: AppServerProcess): Promise<ThreadSearchIndex> {
-  const docs = await loadAllThreadsForSearch(appServer)
-  const docsById = new Map<string, ThreadSearchDocument>(docs.map((doc) => [doc.id, doc]))
-  return { docsById }
 }
 
 export function createCodexBridgeMiddleware(): CodexBridgeMiddleware {
@@ -3402,7 +2531,7 @@ export function createCodexBridgeMiddleware(): CodexBridgeMiddleware {
         const repoUrl = typeof payload?.url === 'string' ? payload.url.trim() : ''
         const basePath = typeof payload?.basePath === 'string' ? payload.basePath.trim() : ''
         try {
-          const clonedPath = await cloneGithubRepositoryIntoBase(repoUrl, basePath)
+          const clonedPath = await cloneGithubRepositoryIntoBase(repoUrl, basePath, { runCommand, persistWorkspaceRoot })
           setJson(res, 200, { data: { path: clonedPath } })
         } catch (error) {
           setJson(res, 400, { error: error instanceof Error ? error.message : 'Failed to clone GitHub repository' })
