+++ src/server/methodCatalog.ts
@@ -1,0 +1,125 @@
+import { spawn } from 'node:child_process'
+import { mkdtemp, readFile } from 'node:fs/promises'
+import { tmpdir } from 'node:os'
+import { join } from 'node:path'
+import { resolveCodexCommand } from '../commandResolution.js'
+import { getSpawnInvocation } from '../utils/commandInvocation.js'
+
+function asRecord(value: unknown): Record<string, unknown> | null {
+  return value !== null && typeof value === 'object' && !Array.isArray(value)
+    ? (value as Record<string, unknown>)
+    : null
+}
+
+export class MethodCatalog {
+  private methodCache: string[] | null = null
+  private notificationCache: string[] | null = null
+
+  private async runGenerateSchemaCommand(outDir: string): Promise<void> {
+    await new Promise<void>((resolve, reject) => {
+      const codexCommand = resolveCodexCommand()
+      if (!codexCommand) {
+        reject(new Error('Codex CLI is not available. Install @openai/codex or set CODEXUI_CODEX_COMMAND.'))
+        return
+      }
+
+      const invocation = getSpawnInvocation(codexCommand, ['app-server', 'generate-json-schema', '--out', outDir])
+      const process = spawn(invocation.command, invocation.args, {
+        stdio: ['ignore', 'ignore', 'pipe'],
+      })
+
+      let stderr = ''
+
+      process.stderr.setEncoding('utf8')
+      process.stderr.on('data', (chunk: string) => {
+        stderr += chunk
+      })
+
+      process.on('error', reject)
+      process.on('exit', (code) => {
+        if (code === 0) {
+          resolve()
+          return
+        }
+
+        reject(new Error(stderr.trim() || `generate-json-schema exited with code ${String(code)}`))
+      })
+    })
+  }
+
+  private extractMethodsFromClientRequest(payload: unknown): string[] {
+    const root = asRecord(payload)
+    const oneOf = Array.isArray(root?.oneOf) ? root.oneOf : []
+    const methods = new Set<string>()
+
+    for (const entry of oneOf) {
+      const row = asRecord(entry)
+      const properties = asRecord(row?.properties)
+      const methodDef = asRecord(properties?.method)
+      const methodEnum = Array.isArray(methodDef?.enum) ? methodDef.enum : []
+
+      for (const item of methodEnum) {
+        if (typeof item === 'string' && item.length > 0) {
+          methods.add(item)
+        }
+      }
+    }
+
+    return Array.from(methods).sort((a, b) => a.localeCompare(b))
+  }
+
+  private extractMethodsFromServerNotification(payload: unknown): string[] {
+    const root = asRecord(payload)
+    const oneOf = Array.isArray(root?.oneOf) ? root.oneOf : []
+    const methods = new Set<string>()
+
+    for (const entry of oneOf) {
+      const row = asRecord(entry)
+      const properties = asRecord(row?.properties)
+      const methodDef = asRecord(properties?.method)
+      const methodEnum = Array.isArray(methodDef?.enum) ? methodDef.enum : []
+
+      for (const item of methodEnum) {
+        if (typeof item === 'string' && item.length > 0) {
+          methods.add(item)
+        }
+      }
+    }
+
+    return Array.from(methods).sort((a, b) => a.localeCompare(b))
+  }
+
+  async listMethods(): Promise<string[]> {
+    if (this.methodCache) {
+      return this.methodCache
+    }
+
+    const outDir = await mkdtemp(join(tmpdir(), 'codex-web-local-schema-'))
+    await this.runGenerateSchemaCommand(outDir)
+
+    const clientRequestPath = join(outDir, 'ClientRequest.json')
+    const raw = await readFile(clientRequestPath, 'utf8')
+    const parsed = JSON.parse(raw) as unknown
+    const methods = this.extractMethodsFromClientRequest(parsed)
+
+    this.methodCache = methods
+    return methods
+  }
+
+  async listNotificationMethods(): Promise<string[]> {
+    if (this.notificationCache) {
+      return this.notificationCache
+    }
+
+    const outDir = await mkdtemp(join(tmpdir(), 'codex-web-local-schema-'))
+    await this.runGenerateSchemaCommand(outDir)
+
+    const serverNotificationPath = join(outDir, 'ServerNotification.json')
+    const raw = await readFile(serverNotificationPath, 'utf8')
+    const parsed = JSON.parse(raw) as unknown
+    const methods = this.extractMethodsFromServerNotification(parsed)
+
+    this.notificationCache = methods
+    return methods
+  }
+}
