+++ src/server/providerModelDiscovery.ts
@@ -1,0 +1,212 @@
+import type { AppServerProcess } from './appServerProcess.js'
+
+function asRecord(value: unknown): Record<string, unknown> | null {
+  return value !== null && typeof value === 'object' && !Array.isArray(value)
+    ? (value as Record<string, unknown>)
+    : null
+}
+
+function readNonEmptyString(value: unknown): string {
+  return typeof value === 'string' && value.trim().length > 0 ? value : ''
+}
+
+function getErrorMessage(payload: unknown, fallback: string): string {
+  if (payload instanceof Error && payload.message.trim().length > 0) return payload.message
+  const record = asRecord(payload)
+  if (!record) return fallback
+  const error = record.error
+  if (typeof error === 'string' && error.length > 0) return error
+  const nestedError = asRecord(error)
+  if (nestedError && typeof nestedError.message === 'string' && nestedError.message.length > 0) return nestedError.message
+  return fallback
+}
+
+export type ProviderModelsResponse = {
+  data: string[]
+  providerId: string
+  source: 'provider'
+}
+
+const PROVIDER_MODELS_FETCH_TIMEOUT_MS = 5_000
+
+function normalizeHeaderValue(value: unknown): string | null {
+  if (typeof value === 'string') {
+    const trimmed = value.trim()
+    return trimmed.length > 0 ? trimmed : null
+  }
+  if (typeof value === 'number' || typeof value === 'boolean') {
+    return String(value)
+  }
+  return null
+}
+
+function normalizeQueryParams(value: unknown): URLSearchParams {
+  const params = new URLSearchParams()
+  const record = asRecord(value)
+  if (!record) return params
+
+  for (const [key, rawValue] of Object.entries(record)) {
+    const normalized = normalizeHeaderValue(rawValue)
+    if (!normalized) continue
+    params.set(key, normalized)
+  }
+
+  return params
+}
+
+function buildProviderModelsUrl(baseUrl: string, queryParams: unknown): URL {
+  const url = new URL(baseUrl)
+  url.pathname = url.pathname.endsWith('/') ? `${url.pathname}models` : `${url.pathname}/models`
+  const extraParams = normalizeQueryParams(queryParams)
+  for (const [key, value] of extraParams.entries()) {
+    url.searchParams.set(key, value)
+  }
+  return url
+}
+
+function normalizeProviderModelsData(payload: unknown): string[] {
+  const record = asRecord(payload)
+  const rows = Array.isArray(record?.data) ? record.data : null
+  if (!rows) {
+    throw new Error('provider /models payload is missing a data array')
+  }
+
+  const ids: string[] = []
+  for (const row of rows) {
+    const entry = asRecord(row)
+    const candidate = readNonEmptyString(entry?.id)
+    if (!candidate || ids.includes(candidate)) continue
+    ids.push(candidate)
+  }
+  return ids
+}
+
+export async function fetchCustomEndpointDefaultModel(baseUrl: string, apiKey: string): Promise<string> {
+  const normalizedBaseUrl = baseUrl.trim()
+  if (!normalizedBaseUrl) return ''
+
+  try {
+    const modelsUrl = buildProviderModelsUrl(normalizedBaseUrl, null)
+    const headers: Record<string, string> = apiKey ? { Authorization: `Bearer ${apiKey}` } : {}
+    const response = await fetch(modelsUrl, { headers, signal: AbortSignal.timeout(PROVIDER_MODELS_FETCH_TIMEOUT_MS) })
+    if (!response.ok) return ''
+    const payload = await response.json() as unknown
+    const modelIds = normalizeProviderModelsData(payload)
+    return modelIds[0] ?? ''
+  } catch {
+    return ''
+  }
+}
+
+export async function readProviderBackedModelIds(appServer: AppServerProcess): Promise<ProviderModelsResponse> {
+  const configPayload = asRecord(await appServer.rpc('config/read', {}))
+  const config = asRecord(configPayload?.config)
+  const providerId = readNonEmptyString(config?.model_provider)
+  if (!providerId) {
+    return { data: [], providerId: '', source: 'provider' }
+  }
+
+  const providers = asRecord(config?.model_providers)
+  const provider = asRecord(providers?.[providerId])
+  if (!provider) {
+    logProviderModelDiscoveryWarning('configured provider is missing from model_providers', { providerId })
+    return { data: [], providerId, source: 'provider' }
+  }
+
+  const wireApi = readNonEmptyString(provider.wire_api)
+  if (wireApi !== 'responses') {
+    return { data: [], providerId, source: 'provider' }
+  }
+
+  const baseUrl = readNonEmptyString(provider.base_url)
+  if (!baseUrl) {
+    logProviderModelDiscoveryWarning('responses provider is missing base_url', { providerId })
+    return { data: [], providerId, source: 'provider' }
+  }
+
+  const headers = new Headers()
+  const configuredHeaders = asRecord(provider.http_headers)
+  if (configuredHeaders) {
+    for (const [key, rawValue] of Object.entries(configuredHeaders)) {
+      const normalized = normalizeHeaderValue(rawValue)
+      if (!normalized) continue
+      headers.set(key, normalized)
+    }
+  }
+
+  const bearerToken = readNonEmptyString(provider.experimental_bearer_token)
+  if (bearerToken && !headers.has('Authorization')) {
+    headers.set('Authorization', `Bearer ${bearerToken}`)
+  }
+
+  const envKey = readNonEmptyString(provider.env_key)
+  const envHttpHeaders = asRecord(provider.env_http_headers)
+  if (envKey || envHttpHeaders) {
+    logProviderModelDiscoveryWarning('provider discovery skipped env-backed auth/header expansion', {
+      providerId,
+      hasEnvKey: Boolean(envKey),
+      hasEnvHttpHeaders: Boolean(envHttpHeaders),
+    })
+  }
+
+  let requestUrl: URL
+  try {
+    requestUrl = buildProviderModelsUrl(baseUrl, provider.query_params)
+  } catch (error) {
+    logProviderModelDiscoveryWarning('provider /models URL was invalid', {
+      providerId,
+      error: getErrorMessage(error, 'invalid url'),
+    })
+    return { data: [], providerId, source: 'provider' }
+  }
+
+  let response: Response
+  try {
+    response = await fetch(requestUrl, {
+      method: 'GET',
+      headers,
+      signal: AbortSignal.timeout(PROVIDER_MODELS_FETCH_TIMEOUT_MS),
+    })
+  } catch (error) {
+    logProviderModelDiscoveryWarning('provider /models request failed', {
+      providerId,
+      error: isTimeoutError(error) ? `request timed out after ${PROVIDER_MODELS_FETCH_TIMEOUT_MS}ms` : getErrorMessage(error, 'network error'),
+    })
+    return { data: [], providerId, source: 'provider' }
+  }
+
+  let payload: unknown = null
+  try {
+    payload = await response.json()
+  } catch (error) {
+    logProviderModelDiscoveryWarning('provider /models response was not valid JSON', {
+      providerId,
+      status: response.status,
+      error: getErrorMessage(error, 'invalid json'),
+    })
+    return { data: [], providerId, source: 'provider' }
+  }
+
+  if (!response.ok) {
+    logProviderModelDiscoveryWarning('provider /models request returned non-2xx', {
+      providerId,
+      status: response.status,
+      statusText: response.statusText,
+    })
+    return { data: [], providerId, source: 'provider' }
+  }
+
+  try {
+    return {
+      data: normalizeProviderModelsData(payload),
+      providerId,
+      source: 'provider',
+    }
+  } catch (error) {
+    logProviderModelDiscoveryWarning('provider /models payload was invalid', {
+      providerId,
+      error: getErrorMessage(error, 'invalid payload'),
+    })
+    return { data: [], providerId, source: 'provider' }
+  }
+}
