+++ src/server/threadSearchIndex.ts
@@ -1,0 +1,197 @@
+import { spawn } from 'node:child_process'
+import type { AppServerProcess } from './appServerProcess.js'
+import { resolveRipgrepCommand } from '../commandResolution.js'
+
+const THREAD_SEARCH_FULL_TEXT_THREAD_LIMIT = 100
+
+function asRecord(value: unknown): Record<string, unknown> | null {
+  return value !== null && typeof value === 'object' && !Array.isArray(value)
+    ? (value as Record<string, unknown>)
+    : null
+}
+
+function readNonEmptyString(value: unknown): string {
+  return typeof value === 'string' && value.trim().length > 0 ? value : ''
+}
+
+export type ThreadSearchDocument = {
+  id: string
+  title: string
+  preview: string
+  messageText: string
+  searchableText: string
+}
+
+export type ThreadSearchIndex = {
+  docsById: Map<string, ThreadSearchDocument>
+}
+
+function extractThreadMessageText(threadReadPayload: unknown): string {
+  const payload = asRecord(threadReadPayload)
+  const thread = asRecord(payload?.thread)
+  const turns = Array.isArray(thread?.turns) ? thread.turns : []
+  const parts: string[] = []
+
+  for (const turn of turns) {
+    const turnRecord = asRecord(turn)
+    const items = Array.isArray(turnRecord?.items) ? turnRecord.items : []
+    for (const item of items) {
+      const itemRecord = asRecord(item)
+      const type = typeof itemRecord?.type === 'string' ? itemRecord.type : ''
+      if (type === 'agentMessage' && typeof itemRecord?.text === 'string' && itemRecord.text.trim().length > 0) {
+        parts.push(itemRecord.text.trim())
+        continue
+      }
+      if (type === 'userMessage') {
+        const content = Array.isArray(itemRecord?.content) ? itemRecord.content : []
+        for (const block of content) {
+          const blockRecord = asRecord(block)
+          if (blockRecord?.type === 'text' && typeof blockRecord.text === 'string' && blockRecord.text.trim().length > 0) {
+            parts.push(blockRecord.text.trim())
+          }
+        }
+        continue
+      }
+      if (type === 'commandExecution') {
+        const command = typeof itemRecord?.command === 'string' ? itemRecord.command.trim() : ''
+        const output = typeof itemRecord?.aggregatedOutput === 'string' ? itemRecord.aggregatedOutput.trim() : ''
+        if (command) parts.push(command)
+        if (output) parts.push(output)
+      }
+    }
+  }
+
+  return parts.join('\n').trim()
+}
+
+export function isExactPhraseMatch(query: string, doc: ThreadSearchDocument): boolean {
+  const q = query.trim().toLowerCase()
+  if (!q) return false
+  return (
+    doc.title.toLowerCase().includes(q) ||
+    doc.preview.toLowerCase().includes(q) ||
+    doc.messageText.toLowerCase().includes(q)
+  )
+}
+
+export function scoreFileCandidate(path: string, query: string): number {
+  if (!query) return 0
+  const lowerPath = path.toLowerCase()
+  const lowerQuery = query.toLowerCase()
+  const baseName = lowerPath.slice(lowerPath.lastIndexOf('/') + 1)
+  if (baseName === lowerQuery) return 0
+  if (baseName.startsWith(lowerQuery)) return 1
+  if (baseName.includes(lowerQuery)) return 2
+  if (lowerPath.includes(`/${lowerQuery}`)) return 3
+  if (lowerPath.includes(lowerQuery)) return 4
+  return 10
+}
+
+export async function listFilesWithRipgrep(cwd: string): Promise<string[]> {
+  return await new Promise<string[]>((resolve, reject) => {
+    const ripgrepCommand = resolveRipgrepCommand()
+    if (!ripgrepCommand) {
+      reject(new Error('ripgrep (rg) is not available'))
+      return
+    }
+
+    const proc = spawn(ripgrepCommand, ['--files', '--hidden', '-g', '!.git', '-g', '!node_modules'], {
+      cwd,
+      env: process.env,
+      stdio: ['ignore', 'pipe', 'pipe'],
+    })
+    let stdout = ''
+    let stderr = ''
+    proc.stdout.on('data', (chunk: Buffer) => { stdout += chunk.toString() })
+    proc.stderr.on('data', (chunk: Buffer) => { stderr += chunk.toString() })
+    proc.on('error', reject)
+    proc.on('close', (code) => {
+      if (code === 0) {
+        const rows = stdout
+          .split(/\r?\n/)
+          .map((line) => line.trim())
+          .filter(Boolean)
+        resolve(rows)
+        return
+      }
+      const details = [stderr.trim(), stdout.trim()].filter(Boolean).join('\n')
+      reject(new Error(details || 'rg --files failed'))
+    })
+  })
+}
+
+async function loadAllThreadsForSearch(appServer: AppServerProcess): Promise<ThreadSearchDocument[]> {
+  const threads: Array<{ id: string; title: string; preview: string }> = []
+  let cursor: string | null = null
+
+  do {
+    const response = asRecord(await appServer.rpc('thread/list', {
+      archived: false,
+      limit: 100,
+      sortKey: 'updated_at',
+      modelProviders: [],
+      cursor,
+    }))
+    const data = Array.isArray(response?.data) ? response.data : []
+    for (const row of data) {
+      const record = asRecord(row)
+      const id = typeof record?.id === 'string' ? record.id : ''
+      if (!id) continue
+      const title = typeof record?.name === 'string' && record.name.trim().length > 0
+        ? record.name.trim()
+        : (typeof record?.preview === 'string' && record.preview.trim().length > 0 ? record.preview.trim() : 'Untitled thread')
+      const preview = typeof record?.preview === 'string' ? record.preview : ''
+      threads.push({ id, title, preview })
+    }
+    cursor = typeof response?.nextCursor === 'string' && response.nextCursor.length > 0 ? response.nextCursor : null
+  } while (cursor)
+
+  const docs: ThreadSearchDocument[] = threads.map((thread) => {
+    const searchableText = [thread.title, thread.preview].filter(Boolean).join('\n')
+    return {
+      id: thread.id,
+      title: thread.title,
+      preview: thread.preview,
+      messageText: '',
+      searchableText,
+    } satisfies ThreadSearchDocument
+  })
+
+  const docsById = new Map<string, ThreadSearchDocument>(docs.map((doc) => [doc.id, doc]))
+  const fullTextThreads = threads.slice(0, THREAD_SEARCH_FULL_TEXT_THREAD_LIMIT)
+  const concurrency = 4
+  for (let offset = 0; offset < fullTextThreads.length; offset += concurrency) {
+    const batch = fullTextThreads.slice(offset, offset + concurrency)
+    const loaded = await Promise.all(batch.map(async (thread) => {
+      try {
+        const readResponse = await appServer.rpc('thread/read', {
+          threadId: thread.id,
+          includeTurns: true,
+        })
+        const messageText = extractThreadMessageText(readResponse)
+        const searchableText = [thread.title, thread.preview, messageText].filter(Boolean).join('\n')
+        return [thread.id, {
+          id: thread.id,
+          title: thread.title,
+          preview: thread.preview,
+          messageText,
+          searchableText,
+        } satisfies ThreadSearchDocument] as const
+      } catch {
+        return null
+      }
+    }))
+    for (const row of loaded) {
+      if (!row) continue
+      docsById.set(row[0], row[1])
+    }
+  }
+
+  return Array.from(docsById.values())
+}
+
+export async function buildThreadSearchIndex(appServer: AppServerProcess): Promise<ThreadSearchIndex> {
+  const docs = await loadAllThreadsForSearch(appServer)
+  const docsById = new Map<string, ThreadSearchDocument>(docs.map((doc) => [doc.id, doc]))
+  return { docsById }
+}
